Fix gate problems: Unittest and CentOS new kernel
This patch is fixing two problems that are currently blocking the gate:

1. Unittests failure

The tests: test_get_deleted_resources_order and
test_get_inconsistent_resources_order were failing because the setUp()
method from its class was being invoked before the
SecurityGroupTestPlugin() was initialized.

That wasn't a problem before, but due to the commit
0c112a9f8997472bc9bc27b11501fc18c2e35dd5 in Neutron the "description"
field from the security groups API was removed in favor of the
standardattrdescription extension, therefor we needed that plugin to be
initialized before.

2. CentOS 7 new kernel version

There's some issue with the new kernel version (3.10.0-862) from CentOS7
which is causing the jobs to fail with an SSH timeout. The problem is
currently on investigation.

For now, let's switch the tempest-ovs-stable job back to ubuntu and
unblock the gate.

Closes-Bug: #1771336
Change-Id: Ia7cba29861bfd3991c81e4af36d8abcf29cac159