Migration tool: Enhance ovn_migration.sh script
This patch adds few commands in the ovn_migration.sh script so that the migration
process can be run in steps. It also adds a new ansible playbook file to configure
'dhcp_renewal_time' in dhcp agent configuration file.

With this patch, the user has to run
  Step 1. ./ovn_migration.sh generate-inventory -> This generates the ansible inventory
           file - hosts_for_migration

  Step 2. ./ovn_migration.sh reduce-mtu -> This reduces the MTU of the pre migration
          VxLAN tenant networks and configures 'dhcp_renewal_time' in dhcp agent
          configuration file in the nodes where DHCP agent docker service is running.

  Step 3. ./ovn_migration.sh start-migration -> This kicks start the migration.

This patch also updates the documentation accordingly.

Change-Id: I27ec5d2e42fa8bb1c9e10d9e45a491e0e448cf9f