Set ovsdb probe interval to 1 minute on the client side
When a failover of ovsdb-server occurs, neither neutron-server nor
ovn-metadata-agent detects the disconnection. Even though this should
be fixed in OVS Python IDL, this patch is adding a default probe
interval from the client side to 60 seconds. This way, when the
connection to the VIP is lost, clients will reconnect and the service
will be back up again.

Change-Id: Idec3344136a85d73969ed8e11d358b66c14d9a4d
Closes-Bug: #1772656
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>