Fix README setup.py check error
The releases repo added a validation for PyPI meta files.
The command line is "python3 setup.py check --restructuredtext --strict".
It seems '.. include::' directive cannot work in this check
and it blocks the release process.

LINKS.rst is simple enough, so this commit move the content of LINKS.rst
into README.rst and doc/source/index.rst, while they are duplicated.

This is a workaround to make Rocky-2 release work and
hopefully we can find a better solution in near future.

Change-Id: I381b4494744c2f069227e83ae02f4c956633f4b9