Fix: Router port always marked as DOWN
Router ports in OVN is never bound to a host given their decentralized
nature. For that reason, the code in Neutron core that change the status
of the ports (router ports are normal ports underneath) was ignoring the
OVN router ports.

This patch is forcing the status of those router ports to ACTIVE in
networking-ovn. A comment was left in the code in case we want to change
the core Neutron approach in the future.

This patch is also enabling two tempest tests that checks for this
router port status which were disabled before.

Change-Id: I5fd18e6e395a783e9591c058b80b63d42b24a3df
Closes-Bug: #1715126