Add the tool to update the networks MTU to entry points
This patch adds the tool to update the networks MTU to the
entry points for networking-ovn. The idea is to have a command
for the migration tool so that we can better productize it.

NOTE: As this patch is adding new binaries, the tripleo job will
fail while trying to build the RPM as the new files are not present
in the spec. This is a chicken-egg problem as we can't merge the
spec file without this patch so I'm making the job non-voting for
now.

Change-Id: I265baabdc80a3df9df07d7c33d850184b432d9d6
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>