Add grenade plugin
This patch is configuring a grenade job which exercises upgrading the
networking-ovn ML2 driver. Note that a previous grenade job already
existed in the experiemental pipeline but was outdated and broken; the
previous job was also experimenting migrating from ML2/OVS to ML2/OVN
but we don't want that anymore since that scenario is now handlded by
ansible and the files can be found at networking-ovn/migration.

A TODO was left about upgrading OVN itself as part of the upgrade
process, this can be added in a follow up patch and will require some
more investigation.

We are also starting as a non-voting job until we are confident enough
to enable it in the gate pipeline.

Change-Id: I3b1593710b42918870d4f0b183ba3616f0ae027f