Register OVN metadata as an agent
This patch registers and lists the OVN metadata agent with Neutron.

Since the OVN Controller / Gateway use the Chassis UUID as its own
agent UUID, the OVN Metadata creates can't re-use that. Instead, this
patch generates a random UUID for each Metadata agent and saves it to the
external_ids' column in the Chassis table. At startup time, the metadata
agent will check if the "neutron:ovn-metadata-id" key is present in the
Chassis external_id column and use that UUID if it's there, otherwise
it will generate a random UUID for it.

This patch uses the same AgentStats mechanism to identify whether the
OVN metadata agent is alive or not by relying on the "nb_cfg" from the
Chassis. The only difference from the OVN controller agent is that the
metadata will save the "nb_cfg" of it's last update into the Chassis
external_ids column.

Conflicts:
	networking_ovn/common/utils.py

Co-Authored-By: Daniel Alvarez <dalvarez@redhat.com>
Closes-Bug: #1719574
Change-Id: Icfc9e76e8371d8f33b264df65ae3a7cbc0ca099e
(cherry picked from commit 53a1c87cbf2de25e4d18696987179f40f6c5d67f)