Support octavia provider driver for OVN
This patch implements the provider driver as per the guidelines
mentioned here [1]. In order to load OVN provider driver,
'enabled_provider_drivers' in octavia.conf should define
'ovn:OVN provider driver'. To create a load balancer with ovn
provider driver, the 'provider': 'ovn' should be passed to the
POST of /v2.0/lbaas/loadbalancers.

OVN Northbound db has a table 'Load_Balancer'. This patch creates
a row for each load balancer created and updates the 'external_ids'
and 'vips' column of this table whenever a pool, member or listener
is created/updated or deleted.

There are few TODOs (which can be follow up patches)
  - Add documentation about how to enable this feature
  - Add a design document
  - Add a CI job. Presently the octavia tempest plugin creates the pools
    and listeners with HTTP protocol. We need to enhance the octavia
    tempest tests to use TCP for provider drivers not supporting HTTP.
    https://review.openstack.org/#/c/589712

[1] - https://docs.openstack.org/octavia/latest/contributor/guides/providers.html

Change-Id: I7b61d799b38ea4d34c594818467fc4e611bc222b