Collect OVN databases logs
This patch is changing the tempest-post-common.yml file to collect
the logs from the OVN database as part of the jobs run.

The patch also enhances the networking-ovn DevStack plugin to allow
configuring the OVN database log levels and sets the level to DEBUG in
the tempest runs to help with troubleshooting problems in the gate.

Change-Id: Ib83524c705d88d0232b03bc3dc320430d21249bd