Change ACL optmization strategy in the spec file
When we first wrote this design document, the strategy was to let
neutron-server work with old Address Sets and  create new Security
Groups modelled as Port Groups in OVN. However, that will make the
code unnecesarily complex and we should still provide a mechanism
to migrate from Address Sets to Port Groups.

This patch is changing the strategy by making sure that the migration
takes place upon neutron-server start if the underlying OVN schema
supports Port Groups. This way, neutron-server doesn't need to deal
with mixed modelled resources. The drawback is that we'll be delaying
a bit the start of neutron-server the first time we need to do the
migration but: 1) this will happen only once. 2) initial testing with
500 ports in 50 different networks with 5 SG rules takes around 8
seconds to migrate so it's reasonable.

Change-Id: I666a5ccc87101b559cd9ee4e95b7399d962e23fe
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>