Generate the hosts inventory for OVN migration based on roles
Prior to this patch, we were relying on overcloud node names to decide
where to run ovn-controller and ovn-dbs. This is a bad idea as names
can be anything.

With TripleO composable roles, the safest option would be to run
ovn-dbs everywhere where neutron_api was running in ML2/OVS and run
ovn-controller where OVS agent was. This patch is figuring that out
through tripleo-ansible-inventory output.

Change-Id: Ic053f5a1ae152d564bf9f0f052ee8c80de786313
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>