Maintenance: Fix router ports
Prior to this patch the maintenance thread would fail to fix a router
port resource upon a failure in OVN to create it.

The maintence thread was calling the add_router_interface() method from
the l3_ovn.py module but, if the router port already existed in Neutron
(but didn't exist in OVN) that method would fail with a PortInUse
exception.

This patch accounts for that problem and if the port router already
exist in Neutron it will only fetch it from the database and continue
the work until it's created in the OVN database as well.

A new functional test has been added which to test this scenario.

Closes-Bug: #1746979
Change-Id: Iae500d17d0efe17f3460dc2f09356675d406abed