migration: handle the new json inventory format and silence greps
Something recently changed how the inventory json file is generated
and now the hosts are available straight away under the
Controller / Compute roles, but we still keep support for the old
form.

We add an extra lookup to get the ip address and include it as
ansible_host.

In this same commit we silence the greps which verify proper
configuration of the deployment scripts to avoid grep findings
being output to the console.

Change-Id: I676a8dfbbac350943d4586bb09fd1e97086091fc