Use neutron's agent_down_time for agent liveness
Instead of returning the immediate status, mimick ML2/OVS's
agent_down_time handling. With this patch, we no longer have an
issue where an agent will appear down because we didn't wait long
enough for ovn-controller to update the Chassis and we don't have
to wait before sending a response to API requests.

Closes-Bug: #1719574
Change-Id: Ib08473e4768c94ad3b0a223b72f1dad1d26769dd