Disable agent extension until it's fully stable
Change I43e8b0a7914e3626dd848b04fdfc54fb0351bae6 introduced
the API extension, but the alive flag is not perfectly stable
and makes tempest fail from time to time.

Terry wilson is working on a fix for this here [1], we will
request him to re-enable the extension within that patch.

[1] https://review.openstack.org/#/c/585027/

Change-Id: Id75a6a97d38dce57c3c8386b3036a16950404469