migration: Do some pre-flight checks before migration
This patch checks the availability of the external networks
and floating ips before generating inventory (so the admin
is alerted as soon as possible) and before starting migration
to make sure that still there are enough floating ips available.

This patch also removes the DVR flag, as the ansible scripts
are independent of the DVR configuration on the system.

Change-Id: Ifca00065ce2457437794e34fd6700394db2c3dc5
(cherry picked from commit 1581234d3909de1456fbf97ed6c1dcb545915dfd)