Fix TripleO guide and ovn.yaml file
Replace the link at step 1 from the TripleO deployment guide with the
actual instructions.

That link was broken didn't contain any deployment steps cause it was
moved to another place at [0]. Since we only needed two steps from that
guide (1. To download the quickstart.sh script and 2. To install the
depedencies) I think we could safely add this steps to our own guide.

For the ovn.yaml we are expliciting adding the configuration option
listing the NTP servers that the undercloud should use as part of it's
deployment.

[0]
https://github.com/openstack/tripleo-quickstart/commit/dcd32355f4e0debd38608de0b47c17550dacd5d2

Change-Id: Ia96e2acc7ae3b93f9224e19345ef459c2898d45d