Fix the migration playbook error
The ovn_migration.yaml playbook uses the tag 'include_role'
which is not allowed. We see the below error.

This patch fixes it

ERROR! 'include_role' is not a valid attribute for a Play

The error appears to have been in '/home/stack/ovn_migration/
tripleo_environment/playbooks/ovn-migration.yml': line 76, column 3, but may
be elsewhere in the file depending on the exact syntax problem.

This patch also fixes another error. When the pre migation VM is created, we
test it by running the command '/sbin/ip'. In some of the cirros images
the right path is '/bin/ip'. So change this to the comand 'date'.

Change-Id: Id8f8e2bbc71fee63f14339bc6041c04f3a177495
Closes-bug: #1788158