Fix IPv6 Neigh Adv issue for router IPs
Presently IPv6 functionality is broken with OVN. When a VM
sends an IPv6 Neighbor Solicitation request for the router IP,
ovn-controller when it responds with Neighbor Adv packet doesn't
set the 'router' bit in the 'flags' field. This results in the
default route getting delted. The fix [1] tried to addressed this
issue. But that fix was not completely tested in the openstack context.
ovn-northd adds Neighbor Adv flows for router IPs both in the logical
switch pipeline and logical router pipeline. The commit [1] addressed
the issue in the logical router pipeline. It should have removed the
IPv6 Neighbor Adv flows from the logical switch pipeline. The patch [2]
has been proposed upstream to remove these flows from the logical
switch pipeline. Until that is fixed in core OVN, this patch addresses
this issue in networking-ovn itself as a temporary fix (with the intention
to backport as patch [2] may not be available for OVS 2.9).

When 'add_router_interface' is called, this patch sets the router ports
peer's Logical_Switch_Port.addresses column to it's mac address instead
of 'router'. If the Logical_Switch_Port.addresses is not set to 'router',
ovn-northd will not add the Neighbor Adv flows in the logical switch
pipeline, there by mitigating this issue.

Once the patch [2] is merged, we can revert these changes.

[1] - https://github.com/openvswitch/ovs/commit/c9756229ed818e4c354d8bfd7c65c656f92fc98b#diff-2c35162acf6ad144624954fdc4c3d9f4
[2] - https://patchwork.ozlabs.org/patch/961565/

Closes-bug: #1788684
(cherry picked from commit 91d26e6573d33558b32c1daea40c401ad43ffc7c)

Conflicts:
	networking_ovn/common/ovn_client.py
Change-Id: I815ce0996f66f6b101c2151a8389a8b3fbe026f6