Fix gateway's router port marked as DOWN
This patch is fixing a problem which left the external/gateway's router
port with status DOWN even after the port was bound to a chassis.

Since the Logical_Router_Port table in OVNNB does not have a "up" column
(like the Logical_Switch_Port table has) the status of those ports can
not be updpated based on it. Therefore, this patch uses the "chassis"
column for that. If the router port has the "chassis" column in OVNNB
database set, the binding and status ACTIVE will be set in Neutron
otherwise it status will be udpated to DOWN (if it's not DOWN yet).

Closes-Bug: #1715126
Change-Id: I91026204eae6df18be47df02f99db2f95c428e2b