Fix bug migrating ACLs to Port Groups
When migrating to Port Groups, we were installing wrong ACLs
matching on old Address Sets. This was because, during migration
time, old Address Sets still existed in NB database and we
were relying on their existence to figure out which one to use.

This patch fixes it by simply picking the auto generated
Address Set if Port Groups is supported in the OVSDB schema.

Change-Id: I706199109a3b7dd5339c90b731f8cb8f04ca4f49
Closes-Bug: #1790118
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>