Fix bug when deleting subnets with revision numbers
We were attempting to delete the revision number for a wrong type
when deleting a subnet. This caused the maintenance task to detect
inconsistencies forever and attempt to fix them with no success.

This patch is fixing the bug and adding converage on functional
tests.

Change-Id: I852a69b81150d2d1afb39bed542ca776548ea114
Closes-Bug: #1790659
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit 52d22da487afb1800c53c3d0b43d5efec036fc24)