Default tempest jobs to Python3
This patch is changing the default Python interpreter in the tempest
jobs to Python3, affecting the following playbooks:

* tempest-dsvm-networking-ovn-ovs-master
* tempest-dsvm-networking-ovn-ovs-relase

The patch also introduces a new python2 tempest job running against the
released version of OVS for testing backward compatibility with the old
python interpreter.

There was a tempest-dsvm-networking-ovn-ovs-master-python3 job
definition that ran in the experiemental gate pipeline that has been
deleted as part of this change as well. We don't need it anymore.

Change-Id: Ifc053c03ad08dd373d7dc0afadf860c848b386c9