Fix Pool creation workflow
Currently it seems that when Listener is created before pool creation,
the pool is not updated correctly in the NBDB leading to loss of VIPs
as the workflow continues.

Also updated the dict to deepcopy to ensure that ovn_lb's external_ids
dict is not directly modified as it seems to be held by Python's memory
context which prevented updation of Pool in the listener's key-value pair
in OVN LB.

Also removed a trivial line which was written by dude.

Closes-Bug: #1792317
Change-Id: Ibb4d6a47fdaede4ccd04125bb55ead3ae31112ab