devstack: fix sample config files to correctly setup gq chassis
When deploying multinode environments with devstack using sample
config files, all nodes were eligible to host a gateway but normally
only one would have external connectivity. This would lead to
some gateways being unreachable.

This patch is setting the 'enable-chassis-as-gw' in the default
sample file and leave it disabled in the compute one. This way,
gateways will be scheduled only on master nodes where external
connectivity is assumed.

Change-Id: I3671b2cd57daa0bc2851a287e6cd88c142b45e41
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>