Fix: agent_alive() KeyError problem
The agent_alive() method may fail with a unhandled KeyError when it
tries to determine whether an agent that has been already deleted from
the AgentStats tracker is alive or not.

This patch handles the problem and in those cases it will just report
the agent as "dead".

Closes-Bug: #1796878
Change-Id: Ieee97721af09474fbbd4f03ab74da409f1e12833
(cherry picked from commit 062a05ade43def985dcd735f13460a973a594939)