Clean MAC_Binding entries when (dis)associating a FIP
When a FIP is assigned to a port, it may happen that it was
previously used by another FIP or gateway port and an entry
in MAC_Binding table exits. If that's the case, the ARP
responder will answer with the wrong MAC address and the FIP
will then be unreachable.

In order to be on the safe side, this patch is deleting all
MAC_Binding entries upon association or disassociation of a
FIP as a workaround. The proper way to fix this should be
making ovn-northd or ovn-controller clearing/updating the
stale entries upon new NAT rules creation.

Details about the bug reported in OVS ML at [0].

[0] https://mail.openvswitch.org/pipermail/ovs-discuss/2018-October/047604.html

Change-Id: Ic04fcbe332adf030ab0ee3aad3a94ad874b18c9a
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit 5181f1106ff839d08152623c25c9a5f6797aa2d7)