Fix update_port() SubnetNotFound race conditional
While in collecting port information in the update_port() method the
code will loop thru the fixed_ips associated with the port and fetch the
subnet the IP belongs to from the Neutron database but, in some cases,
that subnet would have been deleted mid-operation. That would cause the
whole update_port() to error out with a SubnetNotFound exception.

This patch fixes the problem by handling such exception and not
appending that IP to the list of addresses (in OVN DB terms) in case the
subnet has already been deleted.

Change-Id: Ia19324722425184da13a7a0383185486eed62e38
Closes-Bug: #1800846
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>