Fix neutron resource cleanup after ml2/ovs migration
The migration script "delete-neutron-resources" was being
executed on controller-0 instead of the undercloud node regardless
of the defer_to:localhost ansible setting.

In that controller ~/overcloudrc does not exist, so the cleanup
was failing silently.

This commit moves the cleanup to a separate role that we tie
to localhost (undercloud) from ovn-migration.yml

Closes-Bug: #1804194

Change-Id: I05f2411604ba01d170440ac655491a624f98aafc