Support ports belonging to Neutron DHCP agents
Before this patch, both the sync tool and the production code
would assume that 'network:dhcp' owner ports would be for
the OVN metadata service. However, since Neutron DHCP agents
can be deployed as well on OVN environments, we need to support
these ports as well.

This change is doing several things:

- Fixing sync tool so that Neutron DHCP ports are not deleted
  anymore (these ports are owned by DHCP and have a device_id
  as 'dhcp<host_uuid>-<network_id>'). Syncing these ports in
  OVN will no longer result in creating 'localport' ports but
  normal ports to allow non-local DHCP traffic.
- Fixing networking-ovn code to skip such ports when looking
  for the metadata port on a specific network.
- Newly created metadata ports will have a device_id such as
  'ovnmeta-<network_id>'.

Change-Id: I30181e1752f456f30f94818c5350f447c387cbb2
Related-Bug: #1804390
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>