Add "irrelevant-files" section to the rally job definition
This patch is adding the "irrelevant-files" section to the rally job
definition. This would prevent the rally job from running on patches
that which *only* changes files that are irrelevant to it. An example of
patch can be find here: https://review.openstack.org/#/c/622242/

This patch also moves the rally job definition to the
zuul.d/networkin-ovn-jobs.yaml file which consolidates all the Zuul
native jobs at the moment.

Change-Id: I7c0df52e2a49d0aa5164b3f06360788b0f67e293
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>