TrivialFix : Verify Listener Protocol
Currently Octavia sends out the updated difference to provider
driver when LoadBalancer API is called. This means that if the
Listener Protocol is not changed during Listener Update, then
Octavia will pass Unset as Listener Protocol. This patch verifies
if the protocol is actually updated when listener update request
is received by OVN.

Change-Id: Ie946fa3559bafb95fd75d09f705ab2811bb9584b