functional: Do not inherit from SqlTestCaseLight
Inheriting from SqlTestCaseLight in some tests and from
SqlTestCase in some others at the same time may cause errors
in functional tests. According to comments in SqlTestCaseLight
implementation in Neutron, that's intended for unit tests
only. This patch is changing it so that all tests inherit
from the same Sql Base class.

For a long time, we've been hitting failures in gate due to
race conditions in functional tests caused by this issue.

Change-Id: Id9c917f604024f6c82b1ff638ba58bce1f2b306b
Closes-Bug: #1808146
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>