functional: Use different IP addresses for every fake chassis
As of OVS 2.11, the SB schema changed to not allow having two
Encap rows with the same (ip, type) pairs. Some of our
functional tests need more than one chassis and they all get
created with the same (ip, type) values so they'll fail.

This patch changes it so that a new IP address is used every
time we create a fake chassis. It shouldn't have any impact
as ovn-controller is not running so tunnels won't be created
anyways.

Change-Id: I900e847851d10f7faf474b70eaecb47b2b016a24
Closes-Bug: #1810292
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>