Add support for VLAN type networks in networking-ovn
This patch is adding support for VLAN type networks in networking-ovn by
setting the new "reside-on-redirect-chassis" option in the
Logical_Router_Port table for router interfaces that are attached to a
VLAN network.

This option causes the routing for the router interface to be
centralized in the chassis hosting the distributed gateway port. For
more information check the commit message from the OVN core patch at
https://github.com/openvswitch/ovs/commit/85706c34d53d4810f54bec1de662392a3c06a996

As a note, since VLAN wasn't supported before this patch is not offering
any migration for existing router interfaces that were created prior to
this patch.

Change-Id: I419ea318c8d78726b5263eecbb1f70bb34f7d96d
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>