Increase nb_cfg when needed only (reducing the # of transactions)
Prior to this patch, any transaction to NBDB would result in the nb_cfg
from NBDB being increased by 1. That would propagated to the SBDB and
the metadata agent would get an event which would result in yet another
transaction being created to update the Chassis' external_ids (SBDB)
column to inform that the agent was alive.

This patch is changing that behavior to bump the nb_cfg only when
the get_agents() API call is invoked which would reduce the number
of transactions created in the SBDB due to our current health check
mechanism.

Closes-Bug: #1814698
Change-Id: I6f91955444e0601f10f2d5e509d4a4453def5f2b
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>