Target proper Chassis events for agent liveness
Agent stats for the ovn-controller are updated any time there is
any CREATE/UPDATE event happens on a Chassis. This includes things
like external_ids updates for metadata agent liveness. So as long
as the metadata agent is running, ovn-controller agent shows up
as 'alive'. This patch makes sure we only update the ovn-controller
agent status when nb_cfg is incremented.

Change-Id: I44c616b94031d611e43c58c32894ae9d54662bcd
Closes-Bug: #1814975