Improve error handling for ChassisMetadataAgentEvent
The ChassisMetadataAgentEvent class is suppose to handle events which
are triggered when a Chassis entry is created or updated in the OVSDB.
Prior to this patch, it was assumed that the external_ids column of that
entry would contain the "neutron:ovn-metadata-sb-cfg" key in it which is
not always true.

This patch is fixing the problem by handling the case which that key is
not yet present in the Chassis external_ids column.

Change-Id: Ie7f348a70a6f0c67f1822f73e0c559eac46b18bb
Closes-Bug: #1816033
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>
(cherry picked from commit c88d989eee5b4331d9ff024595741439bc10ee15)