Use vhost user client mode, instead of server
When doing the port binding, we need to tell libvirt to be the
vhostuser server so that OVS can connect to the socket as a client.
OVS being the "server" is also deprecated behavior.

Without this change, libvirt and OVS also end up disagreeing about
where the socket lives when booting a VM and using DPDK.

Change-Id: I09a4c31d65ed5daa6e2d5cbf9235343d139bdc77