Do not waste CI resources on changes to migration/*
When a patch only changes files in migration/* we do not need to run
our full CI suit on that patch since we do not have coverage for the
migration scripts upstream.

This patch is not a backport of https://review.openstack.org/#/c/633004/
because the stable/rocky branch is still using legacy zuul files for
most of the jobs.

Change-Id: Iae6f4207b178200c33a09d8b979648de8988278b
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>