Fix OVN provider for Octavia DevStack sample file(s)
This patch is fixing the sample file we distribute for enabling the OVN
provider for Octavia by:

* Enabling the o-da service, this is needed for the driver status update
to work.

* Disabling the amphora driver by default.

* Since the amphora driver is now disabled, we can set the OCTAVIA_NODE
variable to "api" (by default it's "standalone")

* Remove the $GIT_BASE variable from the .sample file because that's not
set and doesn't work.

* Update documentation w.r.t the above changes.

This patch also cleans up the duplicated information about octavia in
the normal .sample file and just point it to the octavia one instead.

Change-Id: I6f1e16c7c1f814b1c899e9ad1fe9bdde1af4a0cd
Signed-off-by: Lucas Alvares Gomes <lucasagomes@gmail.com>