Define a hook to configure Octavia o-hm0 iface
The OVN DHCP responder replies to the dhclient on o-hm0 from Octavia
with unsolicited DHCP options. DHCP client requests specific options
which do not include default gateway, yet DHCP server replies with
Default-Gateway and 169.254.169.254 route. While the route might not be
much critical on Devstack deployments, the addition of a L3 gateway is:
it conflicts with the existing one and thus starts routing traffic via
an undesired device.

This patch disables the gateway on the lb-mgmt-subnet. Be warned that
doing so will prevent amphorae traffic to the outside of the network.
Octavia tempest tests do not depend on that so it is safe to not enable
gateway.

Reference: https://bugzilla.redhat.com/show_bug.cgi?id=1689368

Change-Id: I2c410d11297ef7e55eaead55067f890cfc5537bc