Updates for fixing the Networking-OVN gate
Currently, OVN Gate fails due to different reasons which include:

a) certain tests in Neutron's ML2 Plugin
which check for support of updating Segmentation ID on live provider
networks. Since the same functionality is not present in OVNMechanism
Driver for now, the test need to be skipped to allow the gate to be
functional.

b) networking-ovn-tripleo-c7-s003-oooq-containers job is made
non-voting for the check jobs and commented out in the gate jobs,
because of the existing bug in TripleO:
https://bugs.launchpad.net/tripleo/+bug/1824326

c)This patch removes psycopg2 and postgres since its not required
for functional testing by networking-ovn.

Related-Bug: #1823865
Related-Bug: #1824326
Change-Id: I2675bb765d4f709856da5f91df60619915268a2b