metadata: Resync agent when misconfiguration is detected
In case of migration, the metadata agent is kept running with old
br-migration bridge cached in agent's attribute.  Everything else is
configured correctly but agent still uses old value.  This patch adds a
resync mechanism that's triggered by PortBindingChassisEvent in case
agent fails to use currently cached bridge and the bridge doesn't exist
on the system.

Closes-bug: #1824984
Change-Id: Ia17ad3476e61a729136622b4356ce59234ab023a