Set ostestr concurrency to 0 to fix networking-ovn gate
Currently due to an incompatibility between stestr-2.3.1 and os-testr,
gate jobs are failing in networking-ovn. This is because ostestr passes
--concurrency None to stestr and after [1] merged, stestr
is expecting only integer values. Till [2] is not released in the
updated stestr version, run stestr with --concurrency 0.
--concurrency 0 allows to run as many jobs as there are CPUs in the
system, allowing parallel execution in stestr.
This temporarily fixes the issue with tox and the gate failures
resulting because of this inconsistency. Once [2] is released, this
change can be reverted.

[1]: https://github.com/mtreinish/stestr/commit/c93bc81
[2]: https://github.com/mtreinish/stestr/commit/ac8dac0

Related-Bug: #1827834
Change-Id: I1251f33a2a93b0d99574dc94ca000b1f4eb9be99
(cherry picked from commit b04a2341126becbd7b449be6950cfaae4b01a2e5)